+++ airflow/dags/unzip_and_validate_gtfs_schedule/convert_to_json/METADATA.yml
@@ -1,0 +1,1 @@
+tooltip: "Convert known GTFS files to JSONL"
