+++ airflow/plugins/utils.py
@@ -1,10 +1,18 @@
 import logging
 import os
+import pendulum
 
 import pandas as pd
 from airflow.models import Variable
 from calitp import read_gcfs, save_to_gcfs
+from calitp.storage import (
+    GTFSScheduleFeedExtract,
+    PartitionedGCSArtifact,
+)
 from pandas.errors import EmptyDataError
+from typing import ClassVar, List
+
+SCHEDULE_UNZIPPED_BUCKET = os.environ["CALITP_BUCKET__GTFS_SCHEDULE_UNZIPPED"]
 
 
 def get_auth_secret(auth_secret_key: str) -> str:
@@ -92,3 +100,22 @@ def get_successfully_downloaded_feeds(execution_date):
     status = pd.read_csv(f)
 
     return status[lambda d: d.status == "success"]
+
+
+class GTFSScheduleFeedFile(PartitionedGCSArtifact):
+    bucket: ClassVar[str] = SCHEDULE_UNZIPPED_BUCKET
+    partition_names: ClassVar[List[str]] = GTFSScheduleFeedExtract.partition_names
+    ts: pendulum.DateTime
+    base64_url: str
+    zipfile_path: str
+    original_filename: str
+
+    # if you try to set table directly, you get an error because it "shadows a BaseModel attribute"
+    # so set as a property instead
+    @property
+    def table(self) -> str:
+        return self.filename
+
+    @property
+    def dt(self) -> pendulum.Date:
+        return self.ts.date()
